requirements.txt
albumentations==1.4.8
catboost==1.2.5
efficientnet_pytorch==0.7.1
imblearn==0.0
iterative-stratification==0.1.7
lightgbm==4.2.0
numpy==1.26.4
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.84
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow
polars==1.4.1
pretrainedmodels==0.7.4
protobuf==5.27.3
pyarrow==17.0.0
PyYAML
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
timm==1.0.8
torchmetrics==1.4.0.post0
tqdm==4.66.4
xgboost==2.0.3